chore(repo): move tests between slices
