+++ .circleci/config.yml
@@ -390,7 +390,7 @@ workflows:
           run-unit-tests: 'false'
           run-linting: 'false'
           e2e-test-filter: 'MACOS-Tests'
-          exclude-projects: 'e2e-next,e2e-gatsby,e2e-workspace-integrations,e2e-workspace-core,e2e-react,e2e-web,e2e-angular-extensions,e2e-angular-core,e2e-cli,e2e-nx-plugin,e2e-storybook,e2e-cypress,e2e-node,e2e-linter,e2e-jest'
+          exclude-projects: 'e2e-next,e2e-gatsby,e2e-workspace-create,e2e-workspace-integrations,e2e-workspace-core,e2e-react,e2e-web,e2e-angular-extensions,e2e-angular-core,e2e-cli,e2e-nx-plugin,e2e-storybook,e2e-cypress,e2e-node,e2e-linter,e2e-jest'
           filters:
             branches:
               ignore: master
@@ -426,7 +426,7 @@ workflows:
           run-linting: 'false'
           e2e-test-filter: 'MACOS-Tests'
           run-cypress-tests: 'true'
-          exclude-projects: 'e2e-next,e2e-gatsby,e2e-workspace-integrations,e2e-workspace-core,e2e-react,e2e-web,e2e-angular-extensions,e2e-angular-core,e2e-cli,e2e-nx-plugin,e2e-storybook,e2e-cypress,e2e-node,e2e-linter,e2e-jest'
+          exclude-projects: 'e2e-next,e2e-gatsby,e2e-workspace-create,e2e-workspace-integrations,e2e-workspace-core,e2e-react,e2e-web,e2e-angular-extensions,e2e-angular-core,e2e-cli,e2e-nx-plugin,e2e-storybook,e2e-cypress,e2e-node,e2e-linter,e2e-jest'
           filters:
             branches:
               only: master
