+++ e2e/workspace-integrations/src/workspace-aux-commands.test.ts
@@ -474,480 +474,3 @@ describe('Move Angular Project', () => {
     expect(lib2File).toContain(`extends ${newModule}`);
   });
 });
-
-describe('Move Project', () => {
-  /**
-   * Tries moving a library from ${lib}/data-access -> shared/${lib}/data-access
-   */
-  it('should work for libraries', () => {
-    const proj = newProject();
-    const lib1 = uniq('mylib');
-    const lib2 = uniq('mylib');
-    const lib3 = uniq('mylib');
-    runCLI(`generate @nrwl/workspace:lib ${lib1}/data-access`);
-
-    updateFile(
-      `libs/${lib1}/data-access/src/lib/${lib1}-data-access.ts`,
-      `export function fromLibOne() { console.log('This is completely pointless'); }`
-    );
-
-    updateFile(
-      `libs/${lib1}/data-access/src/index.ts`,
-      `export * from './lib/${lib1}-data-access.ts'`
-    );
-
-    /**
-     * Create a library which imports a class from lib1
-     */
-
-    runCLI(`generate @nrwl/workspace:lib ${lib2}/ui`);
-
-    updateFile(
-      `libs/${lib2}/ui/src/lib/${lib2}-ui.ts`,
-      `import { fromLibOne } from '@${proj}/${lib1}/data-access';
-
-        export const fromLibTwo = () => fromLibOne();`
-    );
-
-    /**
-     * Create a library which has an implicit dependency on lib1
-     */
-
-    runCLI(`generate @nrwl/workspace:lib ${lib3}`);
-    let nxJson = JSON.parse(readFile('nx.json')) as NxJsonConfiguration;
-    nxJson.projects[lib3].implicitDependencies = [`${lib1}-data-access`];
-    updateFile(`nx.json`, JSON.stringify(nxJson));
-
-    /**
-     * Now try to move lib1
-     */
-
-    const moveOutput = runCLI(
-      `generate @nrwl/workspace:move --project ${lib1}-data-access shared/${lib1}/data-access`
-    );
-
-    expect(moveOutput).toContain(`DELETE libs/${lib1}/data-access`);
-    expect(exists(`libs/${lib1}/data-access`)).toBeFalsy();
-
-    const newPath = `libs/shared/${lib1}/data-access`;
-    const newName = `shared-${lib1}-data-access`;
-
-    const readmePath = `${newPath}/README.md`;
-    expect(moveOutput).toContain(`CREATE ${readmePath}`);
-    checkFilesExist(readmePath);
-
-    const jestConfigPath = `${newPath}/jest.config.js`;
-    expect(moveOutput).toContain(`CREATE ${jestConfigPath}`);
-    checkFilesExist(jestConfigPath);
-    const jestConfig = readFile(jestConfigPath);
-    expect(jestConfig).toContain(`displayName: 'shared-${lib1}-data-access'`);
-    expect(jestConfig).toContain(`preset: '../../../../jest.preset.js'`);
-    expect(jestConfig).toContain(`'../../../../coverage/${newPath}'`);
-
-    const tsConfigPath = `${newPath}/tsconfig.json`;
-    expect(moveOutput).toContain(`CREATE ${tsConfigPath}`);
-    checkFilesExist(tsConfigPath);
-
-    const tsConfigLibPath = `${newPath}/tsconfig.lib.json`;
-    expect(moveOutput).toContain(`CREATE ${tsConfigLibPath}`);
-    checkFilesExist(tsConfigLibPath);
-    const tsConfigLib = readJson(tsConfigLibPath);
-    expect(tsConfigLib.compilerOptions.outDir).toEqual(
-      '../../../../dist/out-tsc'
-    );
-
-    const tsConfigSpecPath = `${newPath}/tsconfig.spec.json`;
-    expect(moveOutput).toContain(`CREATE ${tsConfigSpecPath}`);
-    checkFilesExist(tsConfigSpecPath);
-    const tsConfigSpec = readJson(tsConfigSpecPath);
-    expect(tsConfigSpec.compilerOptions.outDir).toEqual(
-      '../../../../dist/out-tsc'
-    );
-
-    const indexPath = `${newPath}/src/index.ts`;
-    expect(moveOutput).toContain(`CREATE ${indexPath}`);
-    checkFilesExist(indexPath);
-
-    const rootClassPath = `${newPath}/src/lib/${lib1}-data-access.ts`;
-    expect(moveOutput).toContain(`CREATE ${rootClassPath}`);
-    checkFilesExist(rootClassPath);
-
-    expect(moveOutput).toContain('UPDATE nx.json');
-    nxJson = JSON.parse(readFile('nx.json')) as NxJsonConfiguration;
-    expect(nxJson.projects[`${lib1}-data-access`]).toBeUndefined();
-    expect(nxJson.projects[newName]).toEqual({
-      tags: [],
-    });
-    expect(nxJson.projects[lib3].implicitDependencies).toEqual([
-      `shared-${lib1}-data-access`,
-    ]);
-
-    expect(moveOutput).toContain('UPDATE tsconfig.base.json');
-    const rootTsConfig = readJson('tsconfig.base.json');
-    expect(
-      rootTsConfig.compilerOptions.paths[`@${proj}/${lib1}/data-access`]
-    ).toBeUndefined();
-    expect(
-      rootTsConfig.compilerOptions.paths[`@${proj}/shared-${lib1}-data-access`]
-    ).toEqual([`libs/shared/${lib1}/data-access/src/index.ts`]);
-
-    expect(moveOutput).toContain(`UPDATE workspace.json`);
-    const workspaceJson = readJson(`workspace.json`);
-    expect(workspaceJson.projects[`${lib1}-data-access`]).toBeUndefined();
-    const project = workspaceJson.projects[newName];
-    expect(project).toBeTruthy();
-    expect(project.root).toBe(newPath);
-    expect(project.sourceRoot).toBe(`${newPath}/src`);
-    expect(project.targets.lint.options.lintFilePatterns).toEqual([
-      `libs/shared/${lib1}/data-access/**/*.ts`,
-    ]);
-
-    /**
-     * Check that the import in lib2 has been updated
-     */
-    const lib2FilePath = `libs/${lib2}/ui/src/lib/${lib2}-ui.ts`;
-    const lib2File = readFile(lib2FilePath);
-    expect(lib2File).toContain(
-      `import { fromLibOne } from '@${proj}/shared-${lib1}-data-access';`
-    );
-  });
-
-  it('should work for libs created with --importPath', () => {
-    const proj = newProject();
-    const importPath = '@wibble/fish';
-    const lib1 = uniq('mylib');
-    const lib2 = uniq('mylib');
-    const lib3 = uniq('mylib');
-    runCLI(
-      `generate @nrwl/workspace:lib ${lib1}/data-access --importPath=${importPath}`
-    );
-
-    updateFile(
-      `libs/${lib1}/data-access/src/lib/${lib1}-data-access.ts`,
-      `export function fromLibOne() { console.log('This is completely pointless'); }`
-    );
-
-    updateFile(
-      `libs/${lib1}/data-access/src/index.ts`,
-      `export * from './lib/${lib1}-data-access.ts'`
-    );
-
-    /**
-     * Create a library which imports a class from lib1
-     */
-
-    runCLI(`generate @nrwl/workspace:lib ${lib2}/ui`);
-
-    updateFile(
-      `libs/${lib2}/ui/src/lib/${lib2}-ui.ts`,
-      `import { fromLibOne } from '${importPath}';
-
-        export const fromLibTwo = () => fromLibOne();`
-    );
-
-    /**
-     * Create a library which has an implicit dependency on lib1
-     */
-
-    runCLI(`generate @nrwl/workspace:lib ${lib3}`);
-    let nxJson = JSON.parse(readFile('nx.json')) as NxJsonConfiguration;
-    nxJson.projects[lib3].implicitDependencies = [`${lib1}-data-access`];
-    updateFile(`nx.json`, JSON.stringify(nxJson));
-
-    /**
-     * Now try to move lib1
-     */
-
-    const moveOutput = runCLI(
-      `generate @nrwl/workspace:move --project ${lib1}-data-access shared/${lib1}/data-access`
-    );
-
-    expect(moveOutput).toContain(`DELETE libs/${lib1}/data-access`);
-    expect(exists(`libs/${lib1}/data-access`)).toBeFalsy();
-
-    const newPath = `libs/shared/${lib1}/data-access`;
-    const newName = `shared-${lib1}-data-access`;
-
-    const readmePath = `${newPath}/README.md`;
-    expect(moveOutput).toContain(`CREATE ${readmePath}`);
-    checkFilesExist(readmePath);
-
-    const jestConfigPath = `${newPath}/jest.config.js`;
-    expect(moveOutput).toContain(`CREATE ${jestConfigPath}`);
-    checkFilesExist(jestConfigPath);
-    const jestConfig = readFile(jestConfigPath);
-    expect(jestConfig).toContain(`displayName: 'shared-${lib1}-data-access'`);
-    expect(jestConfig).toContain(`preset: '../../../../jest.preset.js'`);
-    expect(jestConfig).toContain(`'../../../../coverage/${newPath}'`);
-
-    const tsConfigPath = `${newPath}/tsconfig.json`;
-    expect(moveOutput).toContain(`CREATE ${tsConfigPath}`);
-    checkFilesExist(tsConfigPath);
-
-    const tsConfigLibPath = `${newPath}/tsconfig.lib.json`;
-    expect(moveOutput).toContain(`CREATE ${tsConfigLibPath}`);
-    checkFilesExist(tsConfigLibPath);
-    const tsConfigLib = readJson(tsConfigLibPath);
-    expect(tsConfigLib.compilerOptions.outDir).toEqual(
-      '../../../../dist/out-tsc'
-    );
-
-    const tsConfigSpecPath = `${newPath}/tsconfig.spec.json`;
-    expect(moveOutput).toContain(`CREATE ${tsConfigSpecPath}`);
-    checkFilesExist(tsConfigSpecPath);
-    const tsConfigSpec = readJson(tsConfigSpecPath);
-    expect(tsConfigSpec.compilerOptions.outDir).toEqual(
-      '../../../../dist/out-tsc'
-    );
-
-    const indexPath = `${newPath}/src/index.ts`;
-    expect(moveOutput).toContain(`CREATE ${indexPath}`);
-    checkFilesExist(indexPath);
-
-    const rootClassPath = `${newPath}/src/lib/${lib1}-data-access.ts`;
-    expect(moveOutput).toContain(`CREATE ${rootClassPath}`);
-    checkFilesExist(rootClassPath);
-
-    expect(moveOutput).toContain('UPDATE nx.json');
-    nxJson = JSON.parse(readFile('nx.json')) as NxJsonConfiguration;
-    expect(nxJson.projects[`${lib1}-data-access`]).toBeUndefined();
-    expect(nxJson.projects[newName]).toEqual({
-      tags: [],
-    });
-    expect(nxJson.projects[lib3].implicitDependencies).toEqual([
-      `shared-${lib1}-data-access`,
-    ]);
-
-    expect(moveOutput).toContain('UPDATE tsconfig.base.json');
-    const rootTsConfig = readJson('tsconfig.base.json');
-    expect(
-      rootTsConfig.compilerOptions.paths[`@${proj}/${lib1}/data-access`]
-    ).toBeUndefined();
-    expect(
-      rootTsConfig.compilerOptions.paths[`@${proj}/shared-${lib1}-data-access`]
-    ).toEqual([`libs/shared/${lib1}/data-access/src/index.ts`]);
-
-    expect(moveOutput).toContain(`UPDATE workspace.json`);
-    const workspaceJson = readJson(`workspace.json`);
-    expect(workspaceJson.projects[`${lib1}-data-access`]).toBeUndefined();
-    const project = workspaceJson.projects[newName];
-    expect(project).toBeTruthy();
-    expect(project.root).toBe(newPath);
-    expect(project.sourceRoot).toBe(`${newPath}/src`);
-
-    expect(project.targets.lint.options.lintFilePatterns).toEqual([
-      `libs/shared/${lib1}/data-access/**/*.ts`,
-    ]);
-
-    /**
-     * Check that the import in lib2 has been updated
-     */
-    const lib2FilePath = `libs/${lib2}/ui/src/lib/${lib2}-ui.ts`;
-    const lib2File = readFile(lib2FilePath);
-    expect(lib2File).toContain(
-      `import { fromLibOne } from '@${proj}/shared-${lib1}-data-access';`
-    );
-  });
-
-  it('should work for custom workspace layouts', () => {
-    const proj = newProject();
-    const lib1 = uniq('mylib');
-    const lib2 = uniq('mylib');
-    const lib3 = uniq('mylib');
-
-    let nxJson = readJson('nx.json');
-    nxJson.workspaceLayout = { libsDir: 'packages' };
-    updateFile('nx.json', JSON.stringify(nxJson));
-
-    runCLI(`generate @nrwl/workspace:lib ${lib1}/data-access`);
-
-    updateFile(
-      `packages/${lib1}/data-access/src/lib/${lib1}-data-access.ts`,
-      `export function fromLibOne() { console.log('This is completely pointless'); }`
-    );
-
-    updateFile(
-      `packages/${lib1}/data-access/src/index.ts`,
-      `export * from './lib/${lib1}-data-access.ts'`
-    );
-
-    /**
-     * Create a library which imports a class from lib1
-     */
-
-    runCLI(`generate @nrwl/workspace:lib ${lib2}/ui`);
-
-    updateFile(
-      `packages/${lib2}/ui/src/lib/${lib2}-ui.ts`,
-      `import { fromLibOne } from '@${proj}/${lib1}/data-access';
-
-        export const fromLibTwo = () => fromLibOne();`
-    );
-
-    /**
-     * Create a library which has an implicit dependency on lib1
-     */
-
-    runCLI(`generate @nrwl/workspace:lib ${lib3}`);
-    nxJson = JSON.parse(readFile('nx.json')) as NxJsonConfiguration;
-    nxJson.projects[lib3].implicitDependencies = [`${lib1}-data-access`];
-    updateFile(`nx.json`, JSON.stringify(nxJson));
-
-    /**
-     * Now try to move lib1
-     */
-
-    const moveOutput = runCLI(
-      `generate @nrwl/workspace:move --project ${lib1}-data-access shared/${lib1}/data-access`
-    );
-
-    expect(moveOutput).toContain(`DELETE packages/${lib1}/data-access`);
-    expect(exists(`packages/${lib1}/data-access`)).toBeFalsy();
-
-    const newPath = `packages/shared/${lib1}/data-access`;
-    const newName = `shared-${lib1}-data-access`;
-
-    const readmePath = `${newPath}/README.md`;
-    expect(moveOutput).toContain(`CREATE ${readmePath}`);
-    checkFilesExist(readmePath);
-
-    const jestConfigPath = `${newPath}/jest.config.js`;
-    expect(moveOutput).toContain(`CREATE ${jestConfigPath}`);
-    checkFilesExist(jestConfigPath);
-    const jestConfig = readFile(jestConfigPath);
-    expect(jestConfig).toContain(`displayName: 'shared-${lib1}-data-access'`);
-    expect(jestConfig).toContain(`preset: '../../../../jest.preset.js'`);
-    expect(jestConfig).toContain(`'../../../../coverage/${newPath}'`);
-
-    const tsConfigPath = `${newPath}/tsconfig.json`;
-    expect(moveOutput).toContain(`CREATE ${tsConfigPath}`);
-    checkFilesExist(tsConfigPath);
-
-    const tsConfigLibPath = `${newPath}/tsconfig.lib.json`;
-    expect(moveOutput).toContain(`CREATE ${tsConfigLibPath}`);
-    checkFilesExist(tsConfigLibPath);
-    const tsConfigLib = readJson(tsConfigLibPath);
-    expect(tsConfigLib.compilerOptions.outDir).toEqual(
-      '../../../../dist/out-tsc'
-    );
-
-    const tsConfigSpecPath = `${newPath}/tsconfig.spec.json`;
-    expect(moveOutput).toContain(`CREATE ${tsConfigSpecPath}`);
-    checkFilesExist(tsConfigSpecPath);
-    const tsConfigSpec = readJson(tsConfigSpecPath);
-    expect(tsConfigSpec.compilerOptions.outDir).toEqual(
-      '../../../../dist/out-tsc'
-    );
-
-    const indexPath = `${newPath}/src/index.ts`;
-    expect(moveOutput).toContain(`CREATE ${indexPath}`);
-    checkFilesExist(indexPath);
-
-    const rootClassPath = `${newPath}/src/lib/${lib1}-data-access.ts`;
-    expect(moveOutput).toContain(`CREATE ${rootClassPath}`);
-    checkFilesExist(rootClassPath);
-
-    expect(moveOutput).toContain('UPDATE nx.json');
-    nxJson = JSON.parse(readFile('nx.json')) as NxJsonConfiguration;
-    expect(nxJson.projects[`${lib1}-data-access`]).toBeUndefined();
-    expect(nxJson.projects[newName]).toEqual({
-      tags: [],
-    });
-    expect(nxJson.projects[lib3].implicitDependencies).toEqual([
-      `shared-${lib1}-data-access`,
-    ]);
-
-    expect(moveOutput).toContain('UPDATE tsconfig.base.json');
-    const rootTsConfig = readJson('tsconfig.base.json');
-    expect(
-      rootTsConfig.compilerOptions.paths[`@${proj}/${lib1}/data-access`]
-    ).toBeUndefined();
-    expect(
-      rootTsConfig.compilerOptions.paths[`@${proj}/shared-${lib1}-data-access`]
-    ).toEqual([`packages/shared/${lib1}/data-access/src/index.ts`]);
-
-    expect(moveOutput).toContain(`UPDATE workspace.json`);
-    const workspaceJson = readJson(`workspace.json`);
-    expect(workspaceJson.projects[`${lib1}-data-access`]).toBeUndefined();
-    const project = workspaceJson.projects[newName];
-    expect(project).toBeTruthy();
-    expect(project.root).toBe(newPath);
-    expect(project.sourceRoot).toBe(`${newPath}/src`);
-    expect(project.targets.lint.options.lintFilePatterns).toEqual([
-      `packages/shared/${lib1}/data-access/**/*.ts`,
-    ]);
-
-    /**
-     * Check that the import in lib2 has been updated
-     */
-    const lib2FilePath = `packages/${lib2}/ui/src/lib/${lib2}-ui.ts`;
-    const lib2File = readFile(lib2FilePath);
-    expect(lib2File).toContain(
-      `import { fromLibOne } from '@${proj}/shared-${lib1}-data-access';`
-    );
-
-    nxJson = readJson('nx.json');
-    delete nxJson.workspaceLayout;
-    updateFile('nx.json', JSON.stringify(nxJson));
-  });
-});
-
-describe('Remove Project', () => {
-  /**
-   * Tries creating then deleting a lib
-   */
-  it('should work', () => {
-    newProject();
-    const lib1 = uniq('mylib');
-    const lib2 = uniq('mylib');
-
-    runCLI(`generate @nrwl/workspace:lib ${lib1}`);
-    expect(exists(tmpProjPath(`libs/${lib1}`))).toBeTruthy();
-
-    /**
-     * Create a library which has an implicit dependency on lib1
-     */
-
-    runCLI(`generate @nrwl/workspace:lib ${lib2}`);
-    let nxJson = JSON.parse(readFile('nx.json')) as NxJsonConfiguration;
-    nxJson.projects[lib2].implicitDependencies = [lib1];
-    updateFile(`nx.json`, JSON.stringify(nxJson));
-
-    /**
-     * Try removing the project (should fail)
-     */
-
-    let error;
-    try {
-      runCLI(`generate @nrwl/workspace:remove --project ${lib1}`);
-    } catch (e) {
-      error = e;
-    }
-
-    expect(error).toBeDefined();
-    expect(error.stderr.toString()).toContain(
-      `${lib1} is still depended on by the following projects`
-    );
-    expect(error.stderr.toString()).toContain(lib2);
-
-    /**
-     * Try force removing the project
-     */
-
-    const removeOutputForced = runCLI(
-      `generate @nrwl/workspace:remove --project ${lib1} --forceRemove`
-    );
-
-    expect(removeOutputForced).toContain(`DELETE libs/${lib1}`);
-    expect(exists(tmpProjPath(`libs/${lib1}`))).toBeFalsy();
-
-    expect(removeOutputForced).toContain(`UPDATE nx.json`);
-    nxJson = JSON.parse(readFile('nx.json')) as NxJsonConfiguration;
-    expect(nxJson.projects[`${lib1}`]).toBeUndefined();
-    expect(nxJson.projects[lib2].implicitDependencies).toEqual([]);
-
-    expect(removeOutputForced).toContain(`UPDATE workspace.json`);
-    const workspaceJson = readJson(`workspace.json`);
-    expect(workspaceJson.projects[`${lib1}`]).toBeUndefined();
-  });
-});
