+++ workspace.json
@@ -38,6 +38,7 @@
     "e2e-utils": "e2e/utils",
     "e2e-web": "e2e/web",
     "e2e-workspace-core": "e2e/workspace-core",
+    "e2e-workspace-create": "e2e/workspace-create",
     "e2e-workspace-integrations": "e2e/workspace-integrations",
     "eslint-plugin-nx": "packages/eslint-plugin-nx",
     "express": "packages/express",
